requirements.txt
bleak
configparser
requests
paho-mqtt
influxdb-client
influxdb3-python